Documented and added tests for using "contains in" instead of "equals to" predicate when searching in multivalue fields
